Add some "is test prepared" checks to plom prep commands

+++ plom_server/Preparation/views/classlist_manage.py
@@ -8,10 +8,7 @@ from django.shortcuts import render, redirect
 
 from django_htmx.http import HttpResponseClientRedirect
 
-from ..services import (
-    StagingStudentService,
-    PrenameSettingService,
-)
+from ..services import StagingStudentService, PrenameSettingService, TestPreparedSetting
 
 from Base.base_group_views import ManagerRequiredView
 
@@ -26,6 +23,10 @@ class ClasslistDownloadView(ManagerRequiredView):
 
 class ClasslistView(ManagerRequiredView):
     def get(self, request):
+        # if test is already prepared then redirect to the readonly view
+        if TestPreparedSetting.is_test_prepared():
+            return redirect("prep_classlist_view")
+
         sss = StagingStudentService()
         pss = PrenameSettingService()
 
@@ -40,6 +41,10 @@ class ClasslistView(ManagerRequiredView):
         return render(request, "Preparation/classlist_manage.html", context)
 
     def post(self, request):
+        # if test is already prepared then redirect to the readonly view
+        if TestPreparedSetting.is_test_prepared():
+            return redirect("prep_classlist_view")
+
         context = self.build_context()
         ignore_warnings = request.POST.get("ignoreWarnings", False)
 
@@ -63,6 +68,10 @@ class ClasslistView(ManagerRequiredView):
             return redirect("prep_classlist")
 
     def delete(self, request):
+        # if test is already prepared then redirect to the readonly view
+        if TestPreparedSetting.is_test_prepared():
+            return redirect("prep_classlist_view")
+
         StagingStudentService().remove_all_students()
         return HttpResponseClientRedirect(".")
